Changes from PR feedback
