+++ source/frontend/src/constants/contacts.ts
@@ -1,0 +1,11 @@
+export enum RestrictContactType {
+  ONLY_INDIVIDUALS = 'persons',
+  ONLY_ORGANIZATIONS = 'organizations',
+  ONLY_PIMSUSERS = 'pimsusers',
+}
+
+export const allContactTypes: RestrictContactType[] = [
+  RestrictContactType.ONLY_PIMSUSERS,
+  RestrictContactType.ONLY_INDIVIDUALS,
+  RestrictContactType.ONLY_ORGANIZATIONS,
+];
