+++ source/frontend/src/features/mapSideBar/acquisition/add/AddAcquisitionForm.tsx
@@ -18,8 +18,8 @@ import { UserRegionSelectContainer } from '@/components/common/form/UserRegionSe
 import { Section } from '@/components/common/Section/Section';
 import { SectionField } from '@/components/common/Section/SectionField';
 import { StyledSectionParagraph } from '@/components/common/styles';
-import { RestrictContactType } from '@/components/contact/ContactManagerView/ContactFilterComponent/ContactFilterComponent';
 import * as API from '@/constants/API';
+import { RestrictContactType } from '@/constants/contacts';
 import { useOrganizationRepository } from '@/features/contacts/repositories/useOrganizationRepository';
 import { useProjectProvider } from '@/hooks/repositories/useProjectProvider';
 import { useLookupCodeHelpers } from '@/hooks/useLookupCodeHelpers';
